Update Channels.tsx
This code will allow the user to create a maximum of three channels, display the current channel count, and show an alert when the user attempts to create more channels beyond the limit. In this code, I've integrated react-toastify to display toast notifications for the channel creation limit. The error toast will be shown at the top-right corner of the screen and will automatically close after 5 seconds.